Rename project: Jank
Originally I thought of this project as "the frontend of Ratchet" (see
iamvery/ratchet). The more I've worked on it, it feels like a separate
thing that _depends_ on the Ratchet data format. To that end, it is
helpful to name things in a way that isolates the ideas from one
another. Ratchet is a view template that allows you to use plain HTML
that produces your final view from a Ratchet data structure. Jank is a
front-end tool that subscribes portions of a page to a Ratchet data
stream, updating the view whenever the state changes.
