+++ README.md
@@ -1,6 +1,6 @@
-# ratchet
+# jank
 
 [![Build Status][travis-img]][travis]
 
-[travis-img]: https://travis-ci.org/iamvery/ratchetjs.svg?branch=master
-[travis]: https://travis-ci.org/iamvery/ratchetjs
+[travis-img]: https://travis-ci.org/iamvery/jank.svg?branch=master
+[travis]: https://travis-ci.org/iamvery/jank
