+++ examples/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "ratchet-examples",
-  "description": "ratchet usage examples",
+  "name": "jank-examples",
+  "description": "jank usage examples",
   "private": true,
   "dependencies": {
     "babel-preset-es2015": "^6.6.0",
