+++ lib/jank.js
@@ -1,7 +1,7 @@
 var React = require('react');
 var ReactDOM = require('react-dom');
-var ElementGroups = require('./ratchet/grouping');
-var createComponent = require('./ratchet/component').createComponent;
+var ElementGroups = require('./jank/grouping');
+var createComponent = require('./jank/component').createComponent;
 
 function wrap(elements) {
   var wrapper = document.createElement('div');
@@ -10,7 +10,7 @@ function wrap(elements) {
   return wrapper;
 }
 
-var Ratchet = {
+var Jank = {
   init: function(opts) {
     var data = opts.data;
     var channelInit = opts.channelInit;
@@ -41,4 +41,4 @@ var Ratchet = {
   }
 };
 
-module.exports = Ratchet;
+module.exports = Jank;
