+++ lib/jank/attrs-to-props.js
