+++ lib/jank/component.js
