+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ratchetjs",
+  "name": "jank",
   "version": "0.1.0",
   "description": "Transform view with data.",
   "author": "Jay Hayes (iamvery.com)",
